requirements.txt
flask
pandas
scikit-learn
python-whois